mv use_delete_test => route_hooks_test
